pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crosssim"
version = "0.1.0"
description = "Cross-Simulator Scenario Generation Toolkit for Robotic Navigation"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "CrossSim Contributors"}
]
keywords = ["robotics", "simulation", "navigation", "ros", "gazebo", "scenario-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "PyQt5>=5.15.0",
    "numpy>=1.20.0",
    "pillow>=8.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "opencv-python>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/benpkessler/CrossSim"
Repository = "https://github.com/benpkessler/CrossSim"
Issues = "https://github.com/benpkessler/CrossSim/issues"

[project.scripts]
crosssim-gui = "crosssim.gui.main:main"
crosssim-simple = "crosssim.gui.simple_editor:main"
crosssim-cli = "crosssim.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["crosssim*"]

[tool.setuptools.package-data]
crosssim = ["*.yaml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
